+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog], and this project adheres to
 - Split build instructions into separate file.
 - Include URL to README in usage if manpage is not being built.
 
+### Fixed
+- Ordering between systemd socket file descriptor names and rules.
+
 ## [2.1.3] - 2020-06-01
 
 ### Fixed